Move session to sessions/core/
