+++ packages/core/src/sessions/core/read-session.ts
@@ -1,4 +1,4 @@
-import { Session } from '../session';
+import { Session } from './session';
 import { SessionStore } from '../session-store';
 
 export async function readSession(store: SessionStore, id: string): Promise<Session | null> {
