+++ packages/core/src/sessions/core/session.spec.ts
@@ -2,11 +2,11 @@
 import { deepStrictEqual, notStrictEqual, rejects, strictEqual } from 'assert';
 
 // FoalTS
-import { Config, ConfigTypeError, createService } from '../core';
-import { SESSION_DEFAULT_ABSOLUTE_TIMEOUT, SESSION_DEFAULT_INACTIVITY_TIMEOUT } from './constants';
+import { Config, ConfigTypeError, createService } from '../../core';
+import { SESSION_DEFAULT_ABSOLUTE_TIMEOUT, SESSION_DEFAULT_INACTIVITY_TIMEOUT } from '../constants';
 import { Session } from './session';
-import { SessionState } from './session-state.interface';
-import {  SessionStore } from './session-store';
+import { SessionState } from '../session-state.interface';
+import {  SessionStore } from '../session-store';
 
 function createState(): SessionState {
   return {
