+++ packages/core/src/sessions/core/session.ts
@@ -1,13 +1,13 @@
 // FoalTS
-import { generateToken } from '../common';
-import { Config } from '../core';
+import { generateToken } from '../../common';
+import { Config } from '../../core';
 import {
   SESSION_DEFAULT_ABSOLUTE_TIMEOUT,
   SESSION_DEFAULT_GARBAGE_COLLECTOR_PERIODICITY,
   SESSION_DEFAULT_INACTIVITY_TIMEOUT,
-} from './constants';
-import { SessionState } from './session-state.interface';
-import { SessionStore } from './session-store';
+} from '../constants';
+import { SessionState } from '../session-state.interface';
+import { SessionStore } from '../session-store';
 
 /**
  * Representation of a server/database/file session.
