+++ packages/core/src/sessions/http/set-session-cookie.spec.ts
@@ -11,7 +11,7 @@ import {
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
 } from '../constants';
-import { Session } from '../session';
+import { Session } from '../core/session';
 import { setSessionCookie } from './set-session-cookie';
 
 describe('setSessionCookie', () => {
