+++ packages/core/src/sessions/index.ts
@@ -1,6 +1,6 @@
 export { createSession } from './core/create-session';
 export { readSession } from './core/read-session';
 export { Store, SessionStore, SessionAlreadyExists } from './session-store';
-export { Session } from './session';
+export { Session } from './core/session';
 export { SessionState } from './session-state.interface';
 export { UseSessions } from './use-sessions.hook';
