+++ packages/core/src/sessions/use-sessions.hook.spec.ts
@@ -30,7 +30,7 @@ import {
   SESSION_USER_COOKIE_NAME
 } from './constants';
 import { readSession } from './core/read-session';
-import { Session } from './session';
+import { Session } from './core/session';
 import { SessionState } from './session-state.interface';
 import { SessionStore } from './session-store';
 import { UseSessions } from './use-sessions.hook';
